+++ src/vec.cpp
@@ -108,7 +108,7 @@ mrb_value siren_vec_to_s(mrb_state* mrb, mrb_value self)
   gp_Vec* vec = siren_vec_get(mrb, self);
   char str[128];
   snprintf(str, sizeof(str), "#<Vec:0x%x @x=%f, @y=%f, @z=%f>",
-      (uintptr_t)mrb_cptr(self), vec->X(), vec->Y(), vec->Z());
+      (unsigned int)(uintptr_t)mrb_cptr(self), vec->X(), vec->Y(), vec->Z());
   return mrb_str_new_cstr(mrb, str);
 }
